Struct feroxide::Container

pub struct Container<E: Element> {
    pub contents: Vec<ContainerCompound<E>>,
    pub available_energy: Energy,
}

A container for elements

Fields
contents: Vec<ContainerCompound<E>>

A vector with the contents of this container

available_energy: Energy

The amount of energy available

Methods

impl<E: Element> Container<E>

pub fn react<R: Reaction<E>>(&mut self, reaction: &R) -> bool

Applies given Reaction to Container Removing the elements on the left-hand side and adding the elements on the right-hand side. If there is enough energy for the reaction, that amount will be consumed otherwise the reaction won't occur. Returns if the reaction succeeded

pub fn contains(&self, element: &ContainerCompound<E>) -> bool

Check if the container contains a container compound

pub fn has_elements(&self, elements: &[ContainerCompound<E>]) -> bool

Check if the container has all given elements

pub fn has_enough_compounds_for_reaction(
&self,
reaction: &ElemReaction<E>
) -> bool

Check if the container has all required elements for a reaction to occur NOTE: Ignores electrons

pub fn remove_elements(&mut self, elements: &[ContainerCompound<E>])

Remove given elements from container

pub fn add_elements(&mut self, elements: &[ContainerCompound<E>])

Add given elements to container

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_redox_reactions(&self) -> Vec<(ElemReaction<Ion>, SEP)>

Get all possible redox reactions and their SEP's

pub fn stringify(&self) -> String

Convert container to a nice string for displaying

pub fn ion_from_string(string: &str) -> Option<Container<Ion>>

pub fn molecule_from_string(string: &str) -> Option<Container<Molecule>>