Struct feroxide::ReactionCompound

pub struct ReactionCompound<E: Element> {
    pub element: E,
    pub amount: u16,
}

A reaction compound

Fields
element: E

The element it uses

amount: u16

The amount of moles needed

Methods

impl<E: Element> ReactionCompound<E>

pub fn ion_from_string(symbol: &str) -> Option<ReactionCompound<Ion>>

Convert a string representation of a reaction compound into one

pub fn molecule_from_string(symbol: &str) -> Option<ReactionCompound<Molecule>>

Convert a string representation of a reaction compound into one