Struct feroxide::ReactionSide

pub struct ReactionSide<E: Element> {
    pub compounds: Vec<ReactionCompound<E>>,
}

A side of a reaction

Fields
compounds: Vec<ReactionCompound<E>>

The compounds of this side

Methods

impl<E: Element> ReactionSide<E>

fn ion_from_string(symbol: String) -> Option<ReactionSide<Ion>>

Convert a string representation of a reactionside into one

fn molecule_from_string(symbol: String) -> Option<ReactionSide<Molecule>>

Convert a string representation of a reactionside into one

fn total_charge(&self) -> AtomCharge

Calculate the total charge of this reaction side

fn energy(&self) -> Energy

Calculate the energy this side has

fn total_atoms(&self) -> HashMap<AtomNumber, u16>

Calculate the total amount of atoms this side contains